+++ NEWS.md
@@ -2,6 +2,11 @@
 
 ## New features and enhancements
 
+* MKVToolNix GUI: line edits & combo boxes will now have a "clear text" button
+  appear whenever they're not empty. Part of the implementation of #2654.
+* MKVToolNix GUI: the list of predefined track names is now split up into
+  three lists, one for each track type (audio, video & subtitles). Part of the
+  implementation of #2654.
 * mkvmerge: Ogg reader: if present in the tags of Vorbis and Opus tracks,
   cover art will be converted into attachments.